+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+## [v5.0.0] - 2025-09-17
+
 ### Added
 - New dedicated `update` and `create` commands. Closes [#55](https://github.com/raphaelstolt/lean-package-validator/issues/55).
 
@@ -417,7 +419,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - Initial release.
 
-[Unreleased]: https://github.com/raphaelstolt/lean-package-validator/compare/v4.7.1...HEAD
+[Unreleased]: https://github.com/raphaelstolt/lean-package-validator/compare/v5.0.0...HEAD
+[v5.0.0]: https://github.com/raphaelstolt/lean-package-validator/compare/v4.7.1...v5.0.0
 [v4.7.1]: https://github.com/raphaelstolt/lean-package-validator/compare/v4.7.0...v4.7.1
 [v4.7.0]: https://github.com/raphaelstolt/lean-package-validator/compare/v4.6.0...v4.7.0
 [v4.6.0]: https://github.com/raphaelstolt/lean-package-validator/compare/v4.5.0...v4.6.0